+++ go.mod
@@ -11,6 +11,7 @@ require (
 	github.com/json-iterator/go v1.1.12
 	github.com/klauspost/compress v1.15.9
 	github.com/lestrrat-go/strftime v1.0.6
+	github.com/maxatome/go-testdeep v1.12.0
 	github.com/sirupsen/logrus v1.9.0
 	github.com/stretchr/testify v1.8.0
 	google.golang.org/protobuf v1.28.1
